+++ src/misc/xmlwrapp/document.cpp
@@ -1060,7 +1060,7 @@ namespace {
         try {
             error_messages *p = static_cast<error_messages*>(v);
             if (p) {
-                int     line = xmlLastError.line;
+                long    line = xmlLastError.line;
                 if (line < 0)
                     line = 0;
                 std::string     filename;
@@ -1104,7 +1104,7 @@ namespace {
 
             // handle bug in older versions of libxml2
             if (p) {
-                int     line = xmlLastError.line;
+                long    line = xmlLastError.line;
                 if (line < 0)
                     line = 0;
                 std::string     filename;
